Account for idle time when starting a typing break
If the user has been idle for a while when the typing break starts, it
should be taken into account.  Otherwise, we can get into a situation
where 8 minutes before a 10 minute break starts, the user goes idle,
and returns to the computer to find out they still have a full break
remaining.

However, buffer the elapsed time at intervals of 30s in order to avoid
starting the break at seemingly random times.  This way, we will start
at times such as 5:30 or 6:00 instead of something like 5:43.

GNOME Bug 642788